+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 * Adapt account history on api changes and fixes issue #267
 * Speed up history call, when limit is below 1000
 * Improve unit tests for account history
+* Fix estimate_virtual_op_num, when get_account_history returns an empty entry for an index
 
 0.24.17
 -------